requirements.txt
pyinstaller
streamlit
google-genai
pyautogui
mss
pillow
